[Encode] Set correct weight for B frame
Set correct weight for B frame in impplicit weight mode according to
the distance with it references.

+++ media_driver/agnostic/common/codec/hal/codechal_encode_avc_base.h
@@ -166,7 +166,7 @@ struct EncodeAvcPar
     uint8_t                     SearchControl;
     uint8_t                     BSearchControl;
     uint8_t                     BiMixDisable;
-    uint8_t                     BiWeight;
+    uint32_t                    BiWeight;
     bool                        DisableExtendedMvCostRange;
     bool                        EnableAdaptiveSearch;
     bool                        EnableFBRBypass;
